Fix potential ambiguity while parsing queries involving table functions trino/PR#16014
The keyword COPARTITION is specific to table function invocation. It is not a reserved keyword. In some cases, the word COPARTITION can be ambiguously interpreted: either as table argument alias, or as part of the COPARTITION clause.
To solve the ambiguity, we decided to fail queries using the word "copartition" as table argument alias, while keeping the word non-reserved and available to be used as identifier in other contexts.

Changes adapted from trino/PR#16014
Author: kasiafi

Co-authored-by: kasiafi <30203062+kasiafi@users.noreply.github.com>